Create object_cache.py
this is a file/class where I can cache db obj

This is a cache for the objects, that et pulled out of the database.
This is necessary, to not have duplicate objects with the same id.

Using a cache that maps the ojects to their id has multiple benefits:
 - if you modify the object at any point, all objects with the same id get modified *(copy by reference)*
 - less ram usage